Added a CreateLot route and form with basic functionality
This commit adds tanstack router and includes routes for he stocking
form and a created lot form.  The create lot form has much of the
basic functionality, although it doesn't yes submit the post request
and actually create the Lot, it does capture all of the required
information and has fairly complete validation.