Added initial tests for StockingEventForm
this commit add the initial test file for the stocking event form. The
Tests are currently limited to ensuring that the required field
messages appear if the are empty when the form is submitted. This
commit also adds a checkbox array component to work with react hook
form, and updates the schema to accommodate empty strings that are
submitted by the form in place of empty strings.  Three tests
associated with errors from the checkbox array are currently failing.