Added integration tests for most of the optional stocking event flds
This commit adds unit tests to verify that validation for most of the
non-required fields work as expected.  The form can be submitted
without filling in those field and no associated error messages are
reported.  there are also tests that submit invalid data and verify
that the expected error appears in the document. All tests continue to
pass.