Added integration tests for superrefine validation
This commit adds integration tests for the superrefine validation
fields - those fields that are tested against the values of other
fields. Specifically, lat-lon and publication date.  It also update the
schema for waterbodies and stocking sites to expect an option object
and return the value.  All tests continue to pass.