Added second 'empty' schema for AppliedTag objects
This commit adds a second empty schema for applied tag object that
allows for an object that is completely empty. This schema is used to
validate/allow empty forms.  If any value is populated, all of the
schema validation is triggered. Tests were added to ensure that this
behaviour works as expected. The error messages do not currently
appear on the html form however - that's next.