Added validation and tests to ensure dev stage and fish match
This commit adds a validator function to ensure that the age of the
fish in months matches the selected development stage category.  tests
ensure that the validator works as expected and that an error is
rendered on the page when in-valid combinations of age and dev stage
are selected. All previously passing tests continue to pass.