Added results block to bottom of form
This commit is preparing to post to github pages.  A results block
has been added to the form to hold the values returned from a
successful form submission.  It also updates the data dictionary
placeholder data, adds complete placeholder data for the form, updates
the validator to work as expected with the placeholder data, and adds
a 'reset' button to the form. All tests continue to pass.