Added field for other marks description including tests & validation
This commit adds a field to our form to capture a description of other
marks. The field is required if the other_marks check box is selected,
but optional otherwise. The attribute was also added to the
schema. tests were added to ensure that the validation works as
expected and that an appropriate message is presented to the user if
they try to submit a form with the box checked and the description empty.